@@ -45,26 +45,26 @@
     "url": "https://github.com/legaard/au-schedule/issues"
   },
   "dependencies": {
-    "@angular/animations": "^4.3.6",
-    "@angular/common": "^4.3.6",
-    "@angular/compiler": "^4.3.6",
-    "@angular/core": "^4.3.6",
-    "@angular/forms": "^4.3.6",
-    "@angular/http": "^4.3.6",
-    "@angular/platform-browser": "^4.3.6",
-    "@angular/platform-browser-dynamic": "^4.3.6",
-    "@angular/router": "^4.3.6",
-    "@ngrx/store": "^4.0.2",
+    "@angular/animations": "6.0.9",
+    "@angular/common": "6.0.9",
+    "@angular/compiler": "6.0.9",
+    "@angular/core": "6.0.9",
+    "@angular/forms": "6.0.9",
+    "@angular/http": "6.0.9",
+    "@angular/platform-browser": "6.0.9",
+    "@angular/platform-browser-dynamic": "6.0.9",
+    "@angular/router": "6.0.9",
+    "@ngrx/store": "^6.0.0",
     "core-js": "^2.4.1",
     "node-sass": "^4.5.3",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4"
+    "rxjs": "^6.2.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.3",
-    "@angular/compiler-cli": "^4.3.6",
-    "@angular/language-service": "^4.3.6",
-    "@ngrx/store-devtools": "^4.0.0",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "6.0.9",
+    "@angular/language-service": "6.0.9",
+    "@ngrx/store-devtools": "^6.0.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -80,6 +80,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.8"
   }
 }
